added ajax

+++ baskets/views.py
@@ -35,9 +35,11 @@ def basket_add(request, product_id):
 def basket_remove(request, id):
 
     basket = Basket.objects.get(id=id)
-    if basket.qty>1:
-        basket.qty -= 1
-        basket.save()
-    else:
-        basket.delete()
+    if request.is_ajax():
+        if basket.qty>1:
+             basket.qty -= 1
+             basket.save()
+        else:
+             basket.delete()
+
     return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
